requirements.txt
# ============================================================
# LLM Evaluation Framework - Requirements
# Bachelor's Thesis: Systematic Testing of Diagnostic
# Capabilities of Local Multimodal Language Models Using RAG
# ============================================================

# Core LLM APIs
replicate==1.0.7
google-generativeai>=0.8.0

# PDF Processing
pypdf>=4.0.0
PyPDF2>=3.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Machine Learning & NLP
scikit-learn>=1.3.0
sentence-transformers>=2.2.0  # For semantic similarity (answer_correctness metric)

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP & API
requests>=2.31.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0  # For managing API keys securely
tqdm>=4.66.0  # Progress bars for long experiments
pydantic>=2.0.0  # Data validation

# Development & Testing (optional but recommended)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
